long foutjes eruit =)

+++ Project/src/MixedIngredientType.java
@@ -8,7 +8,7 @@ public class MixedIngredientType extends IngredientType{
     /* *********************************************************
      * CONSTRUCTORS
      * *********************************************************/
-    public MixedIngredientType(String name, int[] standardTemperature) {
+    public MixedIngredientType(String name, long[] standardTemperature) {
         super(name, standardTemperature);
     }
